refactor: extract shared coach-tendency and age-from-birth helpers
Three files had an identical 22-line block for pickOffense / pickDefense
/ toCoachTendencies (roster.repository, depth-chart.publisher, coach-
tendencies.repository), and two files had an identical ageFromBirthDate
(players.repository, roster.repository). Copy-paste with no semantic
variation — changing the null-representation of tendencies or the
leap-year behavior of age would have needed to be touched in each
callsite.

Pull them into shared modules inside the feature folders that own the
concept:

- server/features/coaches/tendency-row.ts exports pickOffense,
  pickDefense, toCoachTendencies, and the CoachTendencyRow type.
- server/features/players/age.ts exports ageFromBirthDate.

All three callers now import instead of redefining. No behavior change;
all 695 server tests pass.
